fix sending ships to target from two planets

+++ Tests/PlanetHealthCalculationTests.cs
@@ -51,7 +51,7 @@ namespace Tests
             return ship;
         }
 
-        private static Planet CreatePlanet(float health = 10F, int radius = 20, int? owner = 0, float x = 0F, float y = 0F)
+        private static Planet CreatePlanet(float health = 10F, float radius = 20, int? owner = 0, float x = 0F, float y = 0F)
         {
             var p = new Planet { Id = _id++, Health = health, Owner = owner, Radius = radius, X = x, Y = y, Neighbors = new int[0] };
             p.SetInboundShips(new List<Ship>());
